+++ scripts/gamemodes/base/shared.lua
@@ -3,6 +3,8 @@ RegisterClientScript("cl_init.lua")
 RegisterClientScript("cl_scoreboard.lua")
 RegisterClientScript("shared.lua")
 
+include("sh_teams.lua")
+
 local gamemode = ScriptedGamemode()
 print("Mode de jeu: " .. gamemode.Name)