+++ scripts/gamemodes/teamdeathmatch.lua
@@ -1,0 +1,6 @@
+return {
+	Name = "teamdeathmatch",
+	Base = "deathmatch",
+	Description = "Team deathmatch, kill the other team!",
+	Properties = {}
+}