+++ scripts/gamemodes/teamdeathmatch/shared.lua
@@ -1,0 +1,4 @@
+RegisterClientScript()
+RegisterClientScript("cl_init.lua")
+
+local gamemode = ScriptedGamemode()