+++ CHANGELOG.md
@@ -2,13 +2,16 @@
 
 ## master (unreleased)
 
-* 
+* Added teams handling in scoreboard and base gamemode
+* Added teamdeatchmatch gamemode
 
 ## Scripting
 
 * Gamemode events are now triggered on the table they used to register the event (this is important when using gamemode inheritance)
 * Added IncomingNetworkPacket:ReadColor() and OutgoingNetworkPacket:WriteColor()
 * Added match.GetPlayerByIndex()
+* Added Scoreboard:UpdatePlayerTeam()
+* Added base gamemode event PlayerTeamUpdate
 * match.BroadcastPacket now has a second optional parameter to disable ready check
 
 ## Beta 1.1